+++ CHANGES.md
@@ -10,6 +10,7 @@
 
 * If you provide a calculation to an output layer that is a different size, don't error, but use the output size. This in theory should not be a change if you've only ever used newer APIs (1.7 and up).
 * If you have a global constant layer on an operator that has union composition (addition) keep with the global layer until the top of the expression, then crop to real data.
+* Be more nuanced in trying to correct GDAL's abuse of RuntimeError for OSError.
 
 ### Fixed